+++ src/resources/views/channel/create.blade.php
@@ -5,9 +5,9 @@
 @stop
 
 @section('content')
-{!! Form::model($channel, ['route' => 'vanilo.admin.channel.store', 'autocomplete' => 'off']) !!}
+{!! Form::model($channel, ['route' => 'vanilo.admin.channel.store', 'autocomplete' => 'off', 'class' => 'row']) !!}
 
-    <x-appshell::card accent="success">
+    <x-appshell::card accent="success" class="col col col-md-8">
         <x-slot:title>{{ __('Channel Details') }}</x-slot:title>
 
         @include('vanilo::channel._form')
@@ -18,6 +18,12 @@
         </x-slot:footer>
 
     </x-appshell::card>
+
+    <div class="col col-md-4">
+        @foreach($sidebarWidgets ?? [] as $widget)
+            @include($widget)
+        @endforeach
+    </div>
 
 {!! Form::close() !!}
 @stop